+++ go.mod
@@ -4,7 +4,7 @@ go 1.20
 
 require (
 	github.com/alexedwards/scs/v2 v2.5.1
-	github.com/go-chi/chi v1.5.4
+	github.com/go-chi/chi/v5 v5.0.10
 	github.com/joho/godotenv v1.5.1
 	github.com/lib/pq v1.10.9
 )
