+++ static/css/main.css
@@ -120,6 +120,7 @@ header {
 
 
 
+
 /* Responsive nav*/
 @media (max-width: 992px) {
     .navbar .links,
