+++ static/css/notification.css
@@ -21,7 +21,7 @@
     align-items: center;
     position: relative;
     transform: translateX(100%);
-    animation: moveleft 0.5s linear forwards;
+    animation: moveleft 0.4s linear forwards;
 }
 
 @keyframes moveleft {
@@ -38,7 +38,7 @@
     width: 100%;
     height: 5px;
     background: green;
-    animation: anim 5s linear forwards;
+    animation: anim 2s linear forwards;
 }
 
 @keyframes anim {
